Simplified the revision provider implementations
